pyproject.toml
[project]
name = "WeddingWebsite"
dependencies = [
    "flask",
    "Flask-SQLAlchemy"
]
version = "0.1.0"
description = "Tim Hewitt and Lauren Zeringue's Wedding Website"

[build-system]
requires = ["flit-core<4"]
build-backend = "flit_core.buildapi"
